Removed webpack alias, for easier storybook integration.
