+++ etsin_finder/frontend/.storybook/webpack.config.js
@@ -1,18 +1,15 @@
-// you can use this file to add your custom webpack plugins, loaders and anything you like.
-// This is just the basic way to add additional webpack configurations.
-// For more information refer the docs: https://storybook.js.org/configurations/custom-webpack-config
+// import shared config files
+const sharedConfig = require('../webpack.config.shared')
+const path = require('path')
+const ExtractTextPlugin = require('extract-text-webpack-plugin')
 
-// IMPORTANT
-// When you add this file, we won't add the default configurations which is similar
-// to "React Create App". This only has babel loader to load JavaScript.
-
-module.exports = {
-  plugins: [
-    // your custom plugins
-  ],
-  module: {
-    rules: [
-      // add your custom rules.
-    ],
+const config = {
+  resolve: {
+    // adds aliases for common locations in app
+    // (../../../utils/checkDataLang => Utils/checkDataLang)
+    alias: sharedConfig.alias,
+    extensions: ['.js', '.jsx', '.css'],
   },
-};
+}
+
+module.exports = config
