+++ etsin_finder/frontend/js/components/dataset/data/accessRights.jsx
@@ -7,8 +7,8 @@ import faUnlock from '@fortawesome/fontawesome-free-solid/faUnlock'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 
-import checkNested from 'Utils/checkNested'
-import checkDataLang from 'Utils/checkDataLang'
+import checkNested from '../../../utils/checkNested'
+import checkDataLang from '../../../utils/checkDataLang'
 import Tooltip from '../../general/tooltip'
 
 export const accessRightsBool = accessRights => {
