+++ etsin_finder/frontend/js/components/dataset/description.jsx
@@ -4,10 +4,10 @@ import styled from 'styled-components'
 import Translate from 'react-translate-component'
 import { inject, observer } from 'mobx-react'
 
-import Accessiblity from 'Stores/view/accessibility'
-import dateFormat from 'Utils/dateFormat'
-import checkNested from 'Utils/checkNested'
-import checkDataLang from 'Utils/checkDataLang'
+import Accessiblity from '../../stores/view/accessibility'
+import dateFormat from '../../utils/dateFormat'
+import checkNested from '../../utils/checkNested'
+import checkDataLang from '../../utils/checkDataLang'
 import Button from '../general/button'
 import Label from '../general/label'
 import AccessRights from './data/accessRights'
