+++ etsin_finder/frontend/js/components/dataset/downloads/dataItem.jsx
@@ -4,9 +4,9 @@ import Translate from 'react-translate-component'
 import translate from 'counterpart'
 import PropTypes from 'prop-types'
 
-import checkDataLang from 'Utils/checkDataLang'
-import sizeParse from 'Utils/sizeParse'
-import checkNested from 'Utils/checkNested'
+import checkDataLang from '../../../utils/checkDataLang'
+import sizeParse from '../../../utils/sizeParse'
+import checkNested from '../../../utils/checkNested'
 import FileIcon from './fileIcon'
 import InfoModal from './infoModal'
 import { InvertedButton, TransparentButton } from '../../general/button'
