+++ etsin_finder/frontend/js/components/dataset/downloads/index.jsx
@@ -1,11 +1,11 @@
 import React, { Component } from 'react'
 import Translate from 'react-translate-component'
 import styled from 'styled-components'
-import DatasetQuery from 'Stores/view/datasetquery'
-import Accessiblity from 'Stores/view/accessibility'
-import checkDataLang from 'Utils/checkDataLang'
-import sizeParse from 'Utils/sizeParse'
-import createTree from 'Utils/createTree'
+import DatasetQuery from '../../../stores/view/datasetquery'
+import Accessiblity from '../../../stores/view/accessibility'
+import checkDataLang from '../../../utils/checkDataLang'
+import sizeParse from '../../../utils/sizeParse'
+import createTree from '../../../utils/createTree'
 import DataItem from './dataItem'
 import { accessRightsBool } from '../data/accessRights'
 import Breadcrumbs from './breadcrumbs'
