+++ etsin_finder/frontend/js/components/dataset/index.jsx
@@ -5,8 +5,8 @@ import styled from 'styled-components'
 import Translate from 'react-translate-component'
 import { inject, observer } from 'mobx-react'
 
-import DatasetQuery from 'Stores/view/datasetquery'
-import Accessibility from 'Stores/view/accessibility'
+import DatasetQuery from '../../stores/view/datasetquery'
+import Accessibility from '../../stores/view/accessibility'
 import Sidebar from './sidebar'
 import Content from './content'
 import ErrorPage from '../errorpage'
