+++ etsin_finder/frontend/js/components/dataset/person.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import Translate from 'react-translate-component'
-import checkDataLang from 'Utils/checkDataLang'
+import checkDataLang from '../../utils/checkDataLang'
 
 export default class Person extends Component {
   constructor(props) {
