+++ etsin_finder/frontend/js/components/dataset/sidebar/index.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { inject, observer } from 'mobx-react'
 
-import checkDataLang from 'Utils/checkDataLang'
-import checkNested from 'Utils/checkNested'
-import dateFormat from 'Utils/dateFormat'
+import checkDataLang from '../../../utils/checkDataLang'
+import checkNested from '../../../utils/checkNested'
+import dateFormat from '../../../utils/dateFormat'
 import SidebarItem from './sidebarItem'
 import Identifier from '../data/identifier'
 import Citation from '../data/citation'
