+++ etsin_finder/frontend/js/components/dataset/versionChanger.jsx
@@ -4,7 +4,7 @@ import { withRouter } from 'react-router-dom'
 import translate from 'counterpart'
 import { withTheme } from 'styled-components'
 
-import Accessibility from 'Stores/view/accessibility'
+import Accessibility from '../../stores/view/accessibility'
 import VersionSelect from './versionselect'
 
 class VersionChanger extends Component {
