+++ etsin_finder/frontend/js/components/search/resultslist/index.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { inject, observer } from 'mobx-react'
 import PropTypes from 'prop-types'
 
-import ElasticQuery from 'Stores/view/elasticquery'
+import ElasticQuery from '../../../stores/view/elasticquery'
 import ListItem from './listItem'
 
 class ResultsList extends Component {
