+++ etsin_finder/frontend/js/layout/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 
-import ErrorBoundary from 'Components/general/errorBoundary'
+import ErrorBoundary from '../components/general/errorBoundary'
 import Header from './header'
 import Footer from './footer'
 import Content from './content'
