+++ etsin_finder/frontend/package.json
@@ -6,18 +6,14 @@
   "author": "Aaron Hakala",
   "license": "ISC",
   "scripts": {
-    "build":
-      "webpack --mode production --config webpack.prod.js --define process.env.NODE_ENV=\"'production'\"",
+    "build": "webpack --mode production --config webpack.prod.js --define process.env.NODE_ENV=\"'production'\"",
     "build-storybook": "build-storybook",
-    "dev-build":
-      "webpack -d --config webpack.config.js --define process.env.NODE_ENV=\"'development'\"",
+    "dev-build": "webpack -d --config webpack.config.js --define process.env.NODE_ENV=\"'development'\"",
     "lint": "eslint js --ext .jsx,.js",
-    "start":
-      "webpack --progress -d --config webpack.config.js --watch  --define process.env.NODE_ENV=\"'development'\"",
+    "start": "webpack --progress -d --config webpack.config.js --watch  --define process.env.NODE_ENV=\"'development'\"",
     "storybook": "start-storybook -p 6006",
     "test": "jest",
-    "watch":
-      "npm update; webpack --progress -d --config webpack.config.js --watch  --define process.env.NODE_ENV=\"'development'\""
+    "watch": "npm update; webpack --progress -d --config webpack.config.js --watch  --define process.env.NODE_ENV=\"'development'\""
   },
   "dependencies": {
     "@fortawesome/fontawesome": "^1.1.4",
@@ -25,7 +21,7 @@
     "@fortawesome/fontawesome-free-solid": "^5.0.10",
     "@fortawesome/react-fontawesome": "^0.0.18",
     "axios": "^0.18.0",
-    "babel-core": "^6.26.0",
+    "babel-core": "^6.26.2",
     "babel-loader": "^7.1.4",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
@@ -39,11 +35,11 @@
     "file-loader": "^1.1.11",
     "formik": "^0.11.11",
     "history": "^4.7.2",
-    "loadable-components": "^1.1.1",
+    "loadable-components": "^1.4.0",
     "mappy-breakpoints": "^0.2.3",
     "mobx": "^4.2.0",
     "mobx-react": "^5.0.0",
-    "node-sass": "^4.7.2",
+    "node-sass": "^4.9.0",
     "nprogress": "^0.2.0",
     "polished": "^1.9.2",
     "prop-types": "^15.6.1",
@@ -51,7 +47,7 @@
     "react-a11y-announcer": "^1.2.0",
     "react-dom": "^16.3.2",
     "react-interpolate-component": "^0.12.0",
-    "react-modal": "^3.4.1",
+    "react-modal": "^3.4.4",
     "react-router-dom": "^4.2.2",
     "react-select": "^1.2.1",
     "react-test-renderer": "^16.3.2",
@@ -59,16 +55,16 @@
     "react-translate-component": "^0.15.1",
     "sass-loader": "^7.0.1",
     "styled-components": "^3.2.6",
-    "uglifyjs-webpack-plugin": "^1.2.4",
-    "webpack": "^4.5.0",
-    "webpack-cli": "^2.0.12",
+    "uglifyjs-webpack-plugin": "^1.2.5",
+    "webpack": "^4.6.0",
+    "webpack-cli": "^2.0.15",
     "webpack-node-externals": "^1.6.0",
     "yup": "^0.24.1"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^3.4.1",
-    "@storybook/addon-links": "^3.4.1",
-    "@storybook/react": "^3.4.1",
+    "@storybook/addon-actions": "^3.4.2",
+    "@storybook/addon-links": "^3.4.2",
+    "@storybook/react": "^4.0.0-alpha.3",
     "axios-mock-adapter": "^1.14.1",
     "babel-eslint": "^8.2.3",
     "babel-jest": "^22.4.3",
@@ -98,7 +94,9 @@
       "Stores/(.*)$": "<rootDir>/js/stores/$1",
       "Utils/(.*)$": "<rootDir>/js/utils/$1"
     },
-    "testPathIgnorePatterns": ["./__tests__/test-setup.js"],
+    "testPathIgnorePatterns": [
+      "./__tests__/test-setup.js"
+    ],
     "testURL": "http://localhost"
   },
   "prettier": {
@@ -116,7 +114,9 @@
             "env",
             {
               "targets": {
-                "browsers": ["last 2 major versions"]
+                "browsers": [
+                  "last 2 major versions"
+                ]
               }
             }
           ],
@@ -129,7 +129,9 @@
         "env",
         {
           "targets": {
-            "browsers": ["last 2 major versions"]
+            "browsers": [
+              "last 2 major versions"
+            ]
           }
         }
       ],
