--- etsin_finder/frontend/webpack.config.shared.js
@@ -1,10 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  alias: {
-    Utils: path.resolve(__dirname, 'js/utils/'),
-    Components: path.resolve(__dirname, 'js/components/'),
-    Routes: path.resolve(__dirname, 'js/routes/'),
-    Stores: path.resolve(__dirname, 'js/stores/'),
-  },
-}
