+++ etsin_finder/frontend/webpack.prod.js
@@ -1,4 +1,3 @@
-const sharedConfig = require('./webpack.config.shared')
 const path = require('path')
 const ExtractTextPlugin = require('extract-text-webpack-plugin')
 const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
@@ -12,7 +11,6 @@ const config = {
     chunkFilename: '[name].bundle.js',
   },
   resolve: {
-    alias: sharedConfig.alias,
     extensions: ['.js', '.jsx', '.css'],
   },
   module: {
